+++ scctool/matchgrabber/custom.py
@@ -5,7 +5,7 @@ import logging
 import requests
 
 # create logger
-module_logger = logging.getLogger('scctool.matchgrabber.custom')
+module_logger = logging.getLogger(__name__)
 
 
 class MatchGrabber(object):