+++ scctool/settings/placeholders.py
@@ -2,7 +2,7 @@
 import logging
 
 # create logger
-module_logger = logging.getLogger('scctool.placeholders')
+module_logger = logging.getLogger(__name__)
 
 
 class PlaceholderList: