+++ scctool/tasks/mapstats.py
@@ -3,7 +3,6 @@ import json
 import logging
 import os
 import time
-from shutil import copyfile
 
 from PyQt5.QtCore import pyqtSignal
 
@@ -11,7 +10,7 @@ import scctool.settings
 from scctool.tasks.liquipedia import LiquipediaGrabber, MapNotFound
 from scctool.tasks.tasksthread import TasksThread
 
-module_logger = logging.getLogger('scctool.tasks.mapstats')
+module_logger = logging.getLogger(__name__)
 
 
 class MapStatsManager:
@@ -32,9 +31,9 @@ class MapStatsManager:
         """Read json data from file."""
         file = scctool.settings.getJsonFile('mapstats')
         if not os.path.exists(file):
-            copyfile(scctool.settings.getResFile('mapstats.json'), file)
+            file = scctool.settings.getResFile('mapstats.json')
         try:
-            with open(scctool.settings.getJsonFile('mapstats'),
+            with open(file,
                       'r',
                       encoding='utf-8-sig') as json_file:
                 data = json.load(json_file)
@@ -74,15 +73,17 @@ class MapStatsManager:
         except Exception as e:
             module_logger.exception("message")
 
-    def selectMap(self, map):
+    def selectMap(self, map, send=True):
         if map in self.__maps.keys():
             self.__current_map = map
             if scctool.settings.config.parser.getboolean(
                     "Mapstats", "mark_played",):
-                played = self.__controller.matchData.wasMapPlayed(map)
+                played = self.__controller.matchControl.\
+                    activeMatch().wasMapPlayed(map)
             else:
                 played = False
-            self.__controller.websocketThread.selectMap(map, played)
+            if send:
+                self.__controller.websocketThread.selectMap(map, played)
 
     def setMapPoolType(self, id):
         self.__mappool = int(id)
@@ -113,7 +114,8 @@ class MapStatsManager:
             for map in self.getCustomMapPool():
                 yield map
         else:
-            for map in self.__controller.matchData.yieldMaps():
+            for map in self.__controller.matchControl.\
+                    activeMatch().yieldMaps():
                 if map and map != "TBD":
                     yield map
 
@@ -182,17 +184,23 @@ class MapStatsManager:
 
     def getData(self):
         out_data = dict()
-        out_data['map'] = self.__current_map
+        if self.__current_map in self.getMapPool():
+            out_data['map'] = self.__current_map
+        else:
+            out_data['map'] = None
         out_data['maps'] = dict()
         for map, data in self.__maps.items():
             if map not in self.getMapPool():
                 continue
             out_data['maps'][map] = dict()
             out_data['maps'][map]['map-name'] = map
+            if out_data['map'] is None:
+                out_data['map'] = map
             if scctool.settings.config.parser.getboolean(
                     "Mapstats", "mark_played",):
                 out_data['maps'][map]['played'] = \
-                    self.__controller.matchData.wasMapPlayed(map)
+                    self.__controller.matchControl.\
+                    activeMatch().wasMapPlayed(map)
             else:
                 out_data['maps'][map]['played'] = False
             for key, item in data.items():