+++ scctool/tasks/textfiles.py
@@ -9,20 +9,20 @@ from scctool.tasks.tasksthread import TasksThread
 
 
 # create logger
-module_logger = logging.getLogger('scctool.tasks.textfiles')
+module_logger = logging.getLogger(__name__)
 
 
 class TextFilesThread(TasksThread):
     """Write streaming data to txt-files if needed."""
 
-    def __init__(self, matchData):
+    def __init__(self, matchControl):
         """Init the thread."""
         super().__init__()
-        self._matchData = matchData
+        self._matchControl = matchControl
         self._q = SetQueue()
         self._available_items = ['team', 'score', 'meta', 'league']
-        self._matchData.dataChanged.connect(self.put)
-        self._matchData.metaChangedSignal.connect(self.put)
+        self._matchControl.dataChanged.connect(self.put)
+        self._matchControl.metaChanged.connect(self.put)
         self.addTask('write', self.__writeTask)
         self.activateTask('write')
 
@@ -51,18 +51,19 @@ class TextFilesThread(TasksThread):
             scctool.settings.casting_data_dir +
             "/teams_vs_long.txt")
         with open(file, mode='w', encoding='utf-8') as f:
-            f.write(self._matchData.getTeamOrPlayer(0) + ' vs ' +
-                    self._matchData.getTeamOrPlayer(1) + "\n")
+            team1 = self._matchControl.activeMatch().getTeamOrPlayer(0)
+            team2 = self._matchControl.activeMatch().getTeamOrPlayer(1)
+            f.write('{} vs {}\n'.format(team1, team2))
 
         file = scctool.settings.getAbsPath(
             scctool.settings.casting_data_dir +
             "/teams_vs_short.txt")
         with open(file, mode='w', encoding='utf-8') as f:
-            f.write(self._matchData.getTeamTag(0) + ' vs ' +
-                    self._matchData.getTeamTag(1) + "\n")
+            f.write(self._matchControl.activeMatch().getTeamTag(0) + ' vs ' +
+                    self._matchControl.activeMatch().getTeamTag(1) + "\n")
 
         for idx in range(2):
-            team = self._matchData.getTeamOrPlayer(idx)
+            team = self._matchControl.activeMatch().getTeamOrPlayer(idx)
             file = scctool.settings.getAbsPath(
                 scctool.settings.casting_data_dir +
                 "/team{}.txt".format(idx + 1))
@@ -73,7 +74,7 @@ class TextFilesThread(TasksThread):
         file = scctool.settings.getAbsPath(
             scctool.settings.casting_data_dir + "/score.txt")
 
-        score = self._matchData.getScore()
+        score = self._matchControl.activeMatch().getScore()
         score_str = str(score[0]) + " - " + str(score[1])
 
         with open(file, mode='w', encoding='utf-8') as f:
@@ -83,7 +84,7 @@ class TextFilesThread(TasksThread):
         file = scctool.settings.getAbsPath(
             scctool.settings.casting_data_dir + "/league.txt")
         with open(file, mode='w', encoding='utf-8') as f:
-            f.write(self._matchData.getLeague())
+            f.write(self._matchControl.activeMatch().getLeague())
 
 
 class SetQueue(queue.Queue):